Issues: #7, #8 - Clean up return code and cache file name
Clean out pseudo-code comments.

Remove always-successful return (regardless of what the sub-module
returned).

Add functionality to get all 3 usernames that are associated with the
PAM authentication (calling user, effective user, and destination user).
Depending on the command that is used, the calling user and destination
user may be the same; however there are also cases (such as `su`) where
they could be different as well. The effective user will likely always
be root, but we gather that as well just to be safe.

Sort the includes at the top of the file. It just makes things a bit
cleaner looking.